requirements.txt
matplotlib
numpy
Pillow
tensorflow